services/backend_for_frontend_ms/requirements.txt
Django==4.0
psycopg2-binary==2.9.2
djangorestframework==3.12.4
markdown==3.3.6
django-oauth-toolkit==1.5.0
